Try extracting output handler codes to a new file to be shared with both transcribe and translate.
